BackEnd/NomadTask/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
django-countries==7.3.2
django-filter==22.1
django-userforeignkey==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
emoji-country-flag==1.3.1
idna==3.3
Markdown==3.4.1
Pillow==9.2.0
pycountry==22.3.5
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
